Do not duplicate KURL parsing in SecurityPolicy::generateReferrer
Parsing KURLs from Strings is expensive, and we were doing it multiple times
in this one method, for the same String.

Profiling on Linux estimates that a second KURL parse accounts for ~1.6%
of time spent in ResourceFetcher::requestResource in
addAdditionalRequestHeaders, but we call this method in all sorts of places
that would appreciate a speedup.

BUG=348655

Review-Url: https://codereview.chromium.org/2413333002
Cr-Original-Commit-Position: refs/heads/master@{#425299}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b71cf510e667b61849daec5641101a39b6034399