[Blink] Display images in the center of the screen
Code was adapted from MediaDocument.cpp.  Additions were
made to let the image be zoomed to full width on both
mobile and desktop.

Had to edit TestExpectations to deal with the virtual tests
failing to work with the command line rebaselining tool,
but some of the suggested changes are included here.

Intent to implement:
https://docs.google.com/document/d/1wJPz4kn-9mEFs_ZEM0hAwwr-YzBdW-5hdX1-jkt81Ng/edit?usp=sharing

Images of proposed changes:
https://drive.google.com/drive/folders/0B8k2SmEN8xS3RHJGSVpNZjVtUUE?usp=sharing

BUG=650456

Review-Url: https://codereview.chromium.org/2376163002
Cr-Original-Commit-Position: refs/heads/master@{#425440}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8512a83031e6531c976f4af85a28dbeef58bd635