Use the pixelSnappedScrollWidth/Height for contentsSize
PaintLayerScrollableArea::contentsSize was using the LayoutUnit::toInt
method to convert to int for the size. However, when sizing a Graphicslayer
from this value, it results in layers that are too narrow in some cases,
giving single pixel gaps between the scrolling content and the scrollbars.

The correct value is the pixelSnappedScrollWidth, that accounts for
sub-pixel position of surrounding elements.

Adds a test that fails in m55 and passes with the change, derived from the
site that the bug was reported for.

R=flackr@chromium.org
BUG=653847
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2416453006
Cr-Original-Commit-Position: refs/heads/master@{#425727}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6c715c7a413e006594b26a2b59783d2d9d01f311