CSP: Fire 'SecurityPolicyViolation' on the offending element.
If we can identify a specific element for a CSP violation (e.g. in the
cases of inline script, style, or event handlers), then target those
elements when firing a 'SecurityPolicyViolation' event.

We'll also need to ensure that the event bubbles so that we don't break
existing collection endpoints that listen on 'Document'.

Review-Url: https://codereview.chromium.org/2421473004
Cr-Original-Commit-Position: refs/heads/master@{#425937}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 357d8e12ff388c450cdc431a3e5865737d84d2e2