[css-grid] Remove a duplicated auto repeat computation for intrinsic sizes
After crrev.com/414446 we can safely remove the temporary fix
(crrev.com/405529) we had to avoid crashes and recursive calls when
computing the auto repeat tracks count for grids with intrinsic sizes.

This change unveiled a couple of bugs in the computation of auto repeat
tracks for intrinsic sizes. The first one is that we were not considering the
existence of definite minimum sizes. In those cases, instead of just returning 1
repetition, we should return the minimum number of repetitions that fulfill the
minimum size requirement.

The second bug was that grids were not properly recomputing the number of
auto repeat tracks when the min|max-content contributions of grid items
changed (whenever the grid container had an intrinsic width).

One Mozilla test had to be updated too because it was wrong.
BUG=621517,633474

Review-Url: https://codereview.chromium.org/2287533002
Cr-Original-Commit-Position: refs/heads/master@{#425958}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 90ea1ed689536a8500f9d92652b2a4a0baa24db6