Don't create MediaKeySystemAccess if called when frame is destroyed
If the frame is closed while the permission prompt is displayed,
the permission prompt is dismissed and this may result in the
requestMediaKeySystemAccess request succeeding. However, the blink
objects may have been cleared, so check if this is the case and simply
reject the request.

BUG=641124
TEST=EME browser_tests pass

Review-Url: https://chromiumcodereview.appspot.com/2420243002
Cr-Original-Commit-Position: refs/heads/master@{#426110}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c82ff5dc0349a1ecb6b14f5b77013f429e691777