Port messages sent by WebIDBFactoryImpl to Mojo.
This is the first of a series of patches that convert the IPC messages
sent by WebIDBFactoryImpl, WebIDBDatabaseImpl and WebIDBCursorImpl to
Mojo messages.

Once all of the IPC sent by IndexedDB and Blob storage are converted to
Mojo they can be moved off of channel-associated interfaces and more
agressive removal of unnecessary layers of abstraction (including
removal of all of the manual thread hopping I added) can be done.

BUG=627484

Review-Url: https://chromiumcodereview.appspot.com/2370643004
Cr-Original-Commit-Position: refs/heads/master@{#426412}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 627e7f73d47910f4255031de55075e2a04b222f6