Remove the redundant layout update from WebSurroundingText::textContent
The only usages of WebSurroundingText are in ContentDetector::FindContentRange and
RenderFrameImpl::OnTextSurroundingSelectionRequest, neither of which performs any
DOM or style change between the initialization of WebSurroundingText and the call
sites of textContent. Hence, the layout update in WebSurroundingText::textContent
is redundant and removed by this patch.

SurroundingText::document is only used by the removed code, so it is also removed.

BUG=657268,590369

Review-Url: https://chromiumcodereview.appspot.com/2430273006
Cr-Original-Commit-Position: refs/heads/master@{#426425}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eadac59641a9a79748d30a49b521e0e79822eee5