Update all non-generic test expectations files for removed tests when updating w3c tests.
This is a follow up to http://crrev.com/2360893002, which made it so that
lines in the generic test expectations file (LayoutTests/TestExpectations)
are updated when tests are removed or renamed in the upstream repo.

It turns out that this should be done for all test expectations files,
not just that one.

BUG=649010

Review-Url: https://chromiumcodereview.appspot.com/2432883003
Cr-Original-Commit-Position: refs/heads/master@{#426509}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 01bcbb2eecb603397dc60604e4bc42c9582ea136