Fix TestingPlatformSupportWithMockScheduler::runForPeriodSeconds
When one calls runForPeriodSeconds(x) you naturally expect the
testing clock to alwats advance by that many seconds.

Unfortunately cc::OrderedSimpleTaskRunner is a strange beast and
the actual time advances to the last delayed task before the
deadline if any.

This patch fixes that and makes the operation of timer fast
forwarding explicit.  It also fixes a bug in
TimerTest.RepeatingTimerDoesNotDrift which was relying on
strange behavior of cc::OrderedSimpleTaskRunner when
setAutoAdvanceNowToPendingTasks is false.

BUG=657517

Review-Url: https://chromiumcodereview.appspot.com/2433323003
Cr-Original-Commit-Position: refs/heads/master@{#426545}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 98561e7b7897f88fbceed76c57cefa990c2cb90a