Fix CFI errors in Oilpan
Oilpan is currently casting an object to a polymorphic type to check if the object has
a vtable or not, but this causes a CFI error if the object doesn't have the polymorphic type.
This CL fixes the issue, follwing the krasin's advice on the bug.

BUG=601193

Review-Url: https://chromiumcodereview.appspot.com/2436123002
Cr-Original-Commit-Position: refs/heads/master@{#426615}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 71b71dabac9102c1f2d48040ca15ca288803571f