bluetooth: Require frame to be connected for characteristic.readValue to succeed
Rejects characteristic.readValue if the frame is not connected or if
the frame disconnects during the operation.

The new adapter has a Health Thermometer service with a measurement interval
characteristic. We use this characteristic because it the only standard
characteristic that supports Read, Write and Notify.

Reconnection tests in follow up patch: http://crrev.com/2438963002

BUG=608538

Review-Url: https://codereview.chromium.org/2443473003
Cr-Original-Commit-Position: refs/heads/master@{#427007}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0d1db684a6570d46b61fa141bc37ac95f3a6876a